requirements.txt
streamlit>=1.37.0
crewai>=0.50.0
chromadb>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.1


